Add changes required by the previous reviewer.
