+++ src/index.css
@@ -123,7 +123,7 @@ body {
 .deletebtn {
   background-color: #fff;
   color: gray;
-  font-size: 1.5rem;
+  font-size: 1.6rem;
   cursor: pointer;
   padding: 0.5rem;
   border: none;
