+++ src/index.html
@@ -4,6 +4,7 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet"href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
   <title>Todo_list</title>
 </head>
 <body>
@@ -17,8 +18,7 @@
       <button type="button" class="add-todo"><i class="fa-solid fa-arrow-turn-down"></i></button>
     </div>
     <div class="list-container">
-      <ul class="list-items">
-      </ul>
+      <ul class="list-items"></ul>
     </div>  
     <button type="button" class="btn d-block">Clear all completed</button>
   </section>
