+++ src/index.js
@@ -1,5 +1,5 @@
 import './index.css';
-import './components/app.js';
+import './components/add-delete.js';
 import '@fortawesome/fontawesome-free/js/fontawesome.js';
 import '@fortawesome/fontawesome-free/js/solid.js';
 import '@fortawesome/fontawesome-free/js/regular.js';
