+++ webpack.common.js
@@ -11,6 +11,17 @@ module.exports = {
         test: /\.css/i,
         use: ['style-loader', 'css-loader'],
       },
+      {
+        test: /\.(png|svg|jpg|jpeg|gif)$/i,
+        use: {
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+            outputPath: 'assests',
+            publicPath: 'assets',
+          },
+        },
+      },
     ],
   },
   plugins: [
